refactor: Extract cipher utils
